requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
changes==0.7.0
charset-normalizer==3.0.1
click==3.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.1.7
django-dynamic-datatb==1.0.10
django-dynamic-model==0.3.0
django-dynamic-models==0.0.5
django-dynamic-models-readonly==0.1.1
django-dynamic-preferences==1.14.0
django-environ==0.9.0
django-jsonfield-1.4.1
django-location-field==2.1.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg2==1.19.4
exceptiongroup==1.1.0
fonttools==4.38.0
future==0.18.3
giturlparse.py==0.0.5
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.0
pandas==1.5.3
path.py==4.4
persisting-theory==1.0
phonenumbers==8.13.6
Pillow==9.4.0
pkginfo==1.9.6
pluggy==1.0.0
plumbum==1.4.2
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
PyYAML==3.13
requests==2.4.3
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semantic-version==2.10.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
twine==1.3.1
uritemplate==0.6
urllib3==1.26.14